requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.2
defusedxml==0.7.1
Django==4.2.7
django-bootstrap-icons==0.8.6
django-fontawesome==1.0
idna==3.4
Pillow==10.1.0
psycopg2-binary==2.9.9
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.7
